Fixed a bug that might freeze penggy while sending IP packets Fixed the abnormal deconnection on inactivity Fixed a bug when penggy doesn't release the lockfile Changed some default paths (under /var) Added a check for directories specified in parameters Added a check for the dialup server type

+++ pengfork/src/p3/misc.c
@@ -55,7 +55,7 @@ p3_sync_buffer (buffer)
       /* Try to find a STOP byte first */
       p = memchr (buffer_start (buffer), P3_STOP, buffer->used);
       if (p)
-        len = (int) p - (int) buffer_start (buffer);
+        len = (int) (p - buffer_start (buffer));
       else
         len = buffer->used;
